+++ src/components/Header/context/patronContext.test.tsx
@@ -6,8 +6,25 @@ import { PatronContext, PatronProvider } from "./patronContext";
 describe("PatronContext", () => {
   const PatronTestComponent = () => {
     const name = React.useContext(PatronContext);
-    return <p>{name}</p>;
+
+    if (!name) {
+      return null;
+    }
+
+    return <p>The patron name is: {name}</p>;
   };
+
+  it("returns an empty string if no value is passed", () => {
+    render(
+      <PatronProvider>
+        <PatronTestComponent />
+      </PatronProvider>
+    );
+
+    expect(
+      screen.queryByText(/The patron name is: Tom Nook/i)
+    ).not.toBeInTheDocument();
+  });
 
   it("passes down the patron name value down to its children", () => {
     render(
@@ -16,6 +33,8 @@ describe("PatronContext", () => {
       </PatronProvider>
     );
 
-    expect(screen.getByText("Tom Nook")).toBeInTheDocument();
+    expect(
+      screen.getByText(/The patron name is: Tom Nook/i)
+    ).toBeInTheDocument();
   });
 });