+++ CHANGELOG.md
@@ -20,6 +20,7 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 - Adds the mobile navigation menu for the `Header` component with `MobileNav` and `MobileNavButton` components.
 - Adds the mobile login menu for the `Header` component.
 - Adds the `useCloseDropDown` hook to manage closing dropdown-like components through clicks outside of it and by pressing the escape key.
+- Adds React context for the `Header`'s patron state data through `PatronContext` and `PatronProvider`.
 
 ## Prerelease